app/requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5
djangorestframework==3.11.2
Markdown==3.2.1
